[PE] Refactored the code
Issues:
1. Pallets are not positioned correctly. The pallet position is placing items wrongly
2. Container.height is not showing in UI
3. Container metrics are not loaded properly in UI
4. Need to reduce the memory load of ContainerOptimizationResult
5. Only 1 container is loaded at the moment

+++ objects/ContainerSummary.py
@@ -12,7 +12,7 @@ class ContainerSummary(PydanticBaseModel):
     routeId: str = Field(default='', description='Route ID. Example: "ROUTE_CHI_NY",')
     origin: str = Field(default='', description='Origin location. Example: "Chicago",')
     destinationInSequence: List[str] = Field(default_factory=list, description='Delivery Destinations in sequence. Example: ["New York"],')
-    totalPallets: float = Field(default=0, description='Total pallets loaded in the container. Example: 6')
+    totalPallets: int = Field(default=0, description='Total pallets loaded in the container. Example: 6')
     totalWeightIn_kg: float = Field(default=0, description='Total weight of units loaded in kg (kilogram). Example: 11200')
     totalVolumeIn_m3: float = Field(default=0, description='Total volume occupancy in {m^3} (meter cubic). Example: 864')
